Changed the ids' data type as Long

+++ app/src/main/java/com/atakanmadanoglu/musicapp/data/datasource/local/LocalMusicDataSource.kt
@@ -6,5 +6,5 @@ import kotlinx.coroutines.flow.Flow
 interface LocalMusicDataSource {
     fun getMusics(): Flow<List<FavoriteTrackEntity>>
     suspend fun addMusic(favoriteTrackEntity: FavoriteTrackEntity)
-    suspend fun deleteMusicById(musicId: Int)
+    suspend fun deleteMusicById(musicId: Long)
 }
