+++ app/src/main/java/com/atakanmadanoglu/musicapp/presentation/music_categories/MusicCategoriesScreen.kt
@@ -43,7 +43,7 @@ import com.atakanmadanoglu.musicapp.presentation.navigation.Screen
 fun MusicCategoriesRoute(
     currentRoute: String,
     onBottomNavItemClicked: (route: String) -> Unit,
-    onCardClicked: (itemId: Int, itemName: String) -> Unit,
+    onCardClicked: (itemId: Long, itemName: String) -> Unit,
     musicCategoryViewModel: MusicCategoriesViewModel = hiltViewModel()
 ) {
     val uiState by musicCategoryViewModel.musicCategoryUiState.collectAsStateWithLifecycle()
@@ -64,7 +64,7 @@ private fun MusicCategoriesScreen(
     modifier: Modifier = Modifier,
     genres: List<GenreUI>,
     onBottomNavItemClicked: (route: String) -> Unit,
-    onCardClicked: (itemId: Int, itemName: String) -> Unit,
+    onCardClicked: (itemId: Long, itemName: String) -> Unit,
     currentRoute: String
 ) {
     Scaffold(
@@ -142,7 +142,7 @@ fun BottomNavigationBar(
 fun MusicCategoryVerticalGridList(
     genres: List<GenreUI>,
     contentPaddingValues: PaddingValues,
-    onCardClicked: (itemId: Int, itemName: String) -> Unit
+    onCardClicked: (itemId: Long, itemName: String) -> Unit
 ) {
     LazyVerticalGrid(
         contentPadding = contentPaddingValues,
